decouple UI from container
